requirements.txt
python-telegram-bot
pandas
json5
requests
python-dotenv
bcrypt
ttkbootstrap
Pillow
matplotlib
schedule
tk
